move shell_execute to //:common
