+++ BUILD
@@ -32,3 +32,9 @@ platform(
         }
         """,
 )
+
+py_library(
+    name = "common",
+    srcs = ["common.py"],
+    visibility = ["//visibility:public"]
+)
